feat: rename task_expansion_tile.dart
