+++ apps/tasks/lib/components/patient_bottom_sheet.dart
@@ -3,7 +3,7 @@ import 'package:helpwave_localization/localization.dart';
 import 'package:helpwave_theme/constants.dart';
 import 'package:helpwave_theme/theme.dart';
 import 'package:provider/provider.dart';
-import 'package:tasks/components/TaskExpansionTile.dart';
+import 'package:tasks/components/task_expansion_tile.dart';
 import 'package:tasks/dataclasses/task.dart';
 
 class PatientBottomSheet extends StatefulWidget {
